Add more inclusive gender options
And allow it to be updated from within the client's profile.

This doesn't appear to be used anywhere in our default modules, but it's at least nice to give more inclusive options and allow it to be update by the client, for if it is ever used by a module.