When you press the "Mark as paid" Button on the invoice, you can now enter a transaction id. If the Payment gateway of the Invoice is "Custom" and it's enabled, the transaction ID will be used to add funds in the amount of the invoice to the customer.
Then, it will be marked as paid, just like with any other payment provider.
However, it won't mark the invoice as paid anymore if the payment provider isn't "Custom".

This button has not been touched since 9 Years, so I doubt this will have a high impact, but if you think the behaviour should be different then what it is, lmk.