Fix: emails can change the page layout / styling
This was a PITA, but I got it to (mostly) work quite well.
The full list does as it should, but the individual email page didn't work correctly with the autosizing so that's just hardcoded in size.

I don't think most will use that page over the complete list, but even if they do an an email is too long a scroll bar will be visible so it's good enough till we find a better solution