Refactor error handling in eNom.php to throw Registrar_Exception
Upon reviewing the AdapterAbstract class, noticed that all functions in eNom.php are currently set to return false in case of an error. To adhere to proper error handling practices and maintain consistency with the AdapterAbstract class, refactored the code to throw a Registrar_Exception instead of returning false. This change ensures that errors are handled more robustly and provides clearer feedback when something goes wrong.