Improve test isolation by using temp directories and files (#3120)
* Improve test isolation by using temp directories and files

Updated legacy ServiceTest classes for Extension, System, and Theme modules to use temporary directories and files for better test isolation and cleanup. This prevents side effects on the filesystem and ensures tests do not interfere with existing data or configuration.

* Improve test isolation by using temp directories and files

Updated legacy ServiceTest classes for Extension, System, and Theme modules to use temporary directories and files for better test isolation and cleanup. This prevents side effects on the filesystem and ensures tests do not interfere with existing data or configuration.